{{cookiecutter.project_name}}/pyproject.toml
[tool.poetry]
name = "{{cookiecuter.project_name}}"
version = "0.1.0"
description = "{{cookiecutter.description}}"
authors = ["{{cookiecutter.name}} <{{cookiecutter.email}}>"]

[tool.poetry.dependencies]
python = "^3.8"
pydantic = "^1.8.2"
pytorch-lightning = "^{{cookiecutter.pl_version}}"
lescode = "^0.2.11"

[tool.poetry.dev-dependencies]
class-registry = "^2.1.2"
typer = "^0.3.2"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
